Add unit tests for parsers, install pod to catch NSExpression and refactor/reorganise configuration related classes (#55)
* Fix merge issues

* Fix game parameters bug

* Reorganise and refactor configuration related classes

* Add tests for game parameters parsers

* Add NSException catching for NSExpression using newly installed pod

* Modify tests

+++ Relentless/Relentless/Game Logic/Events/EventGenerator.swift
@@ -10,14 +10,14 @@ import Foundation
 
 class EventGenerator {
 
-    var probabilityOfEvent: Float
+    var probabilityOfEvent: Double
 
-    init(probabilityOfEvent: Float) {
+    init(probabilityOfEvent: Double) {
         self.probabilityOfEvent = probabilityOfEvent
     }
 
     func generate() -> EventType? {
-        let randomNumber = Float.random(in: 0...1)
+        let randomNumber = Double.random(in: 0...1)
         guard randomNumber <= probabilityOfEvent else {
             return nil
         }
